Fix warning about generated provider
