+++ build.yaml
@@ -12,6 +12,7 @@ targets:
           to_json: false
       riverpod_generator:
         generate_for:
+          - lib/src/localizations.dart
           - lib/src/model/**/*.dart
           - lib/src/network/*.dart
           - lib/src/db/*.dart
