+++ lib/src/utils/l10n.dart
@@ -1,51 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:lichess_mobile/l10n/l10n.dart';
-
-typedef ActiveLocalizations = ({
-  Locale locale,
-  AppLocalizations strings,
-});
-
-/// A provider that exposes the current locale and localized strings.
-final l10nProvider =
-    NotifierProvider<L10nNotifier, ActiveLocalizations>(L10nNotifier.new);
-
-class L10nNotifier extends Notifier<ActiveLocalizations> {
-  @override
-  ActiveLocalizations build() {
-    final observer = _LocaleObserver((locales) {
-      _update();
-    });
-    final binding = WidgetsBinding.instance;
-    binding.addObserver(observer);
-    ref.onDispose(() => binding.removeObserver(observer));
-
-    return _getLocale();
-  }
-
-  void _update() {
-    state = _getLocale();
-  }
-
-  ActiveLocalizations _getLocale() {
-    final locale = WidgetsBinding.instance.platformDispatcher.locale;
-    return (
-      locale: locale,
-      strings: lookupAppLocalizations(locale),
-    );
-  }
-}
-
-/// observer used to notify the caller when the locale changes
-class _LocaleObserver extends WidgetsBindingObserver {
-  _LocaleObserver(this._didChangeLocales);
-  final void Function(List<Locale>? locales) _didChangeLocales;
-  @override
-  void didChangeLocales(List<Locale>? locales) {
-    _didChangeLocales(locales);
-  }
-}
 
 /// Returns a localized string with a single placeholder replaced by a widget.
 ///
